Better handling on input
The input datasets now only have to be specified once, everything else
is handled by adaptable loops. Empty datasets can now also be safely
included: the program will print a message informing the user if a
datset is empty and safely continue if it encounters one.

Squashed commit of the following:

commit d5b7a5c2b1f20559fad6ac32df1ca5b75c31da1a
Author: Corin <lunik1@users.noreply.github.com>
Date:   Tue Feb 21 16:30:30 2017 +0000

    Handle empty trees gracefully

commit 982300939c0c8b6678f624f6d97756aa45679d09
Author: Corin <lunik1@users.noreply.github.com>
Date:   Tue Feb 21 15:53:52 2017 +0000

    Fewer loops

commit 9f6d535bf7dcb2a282364146370baf3f0dbd6005
Author: Corin <lunik1@users.noreply.github.com>
Date:   Tue Feb 21 15:02:07 2017 +0000

    Load trees wth loops